requirements.txt
Django==4.2.5
gunicorn==20.0.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
mysqlclient==2.0.3